begin writing tests
