+++ .eslintignore
@@ -1,1 +1,2 @@
 /apidoc
+/node_modules
