+++ config/config.json
@@ -8,7 +8,7 @@
   },
   "test": {
     "username": "root",
-    "password": null,
+    "password": "1234",
     "database": "testdb",
     "host": "localhost",
     "dialect": "mysql",
