+++ views/layouts/main.handlebars
@@ -24,6 +24,7 @@
   {{{body}}}
 
 <script src="js/animate.js"></script>
+<script src="js/index.js"></script>
 </body>
